+++ app/Http/Controllers/Backend/EquipmentTypeController.php
@@ -6,8 +6,10 @@ use App\Http\Controllers\Controller;
 use App\Models\EquipmentType;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
-use Illuminate\Support\Facades\Storage;
 use Intervention\Image\Facades\Image;
+use App\Models\ItemLocations;
+use App\Models\Locations;
+use Torann\GeoIP\Location;
 
 class EquipmentTypeController extends Controller
 {
@@ -18,8 +20,8 @@ class EquipmentTypeController extends Controller
      */
     public function index()
     {
-        $equipmentTypes = EquipmentType::paginate(12);
-        return view('backend.equipment.types.index', compact('equipmentTypes'));
+        //$equipmentTypes = EquipmentType::orderBy('id', 'asc')->paginate(16);
+        return view('backend.equipment.types.index');
     }
 
     /**
@@ -29,7 +31,7 @@ class EquipmentTypeController extends Controller
      */
     public function create()
     {
-        $types = EquipmentType::pluck('title', 'id');
+        $types =  EquipmentType::getFullTypeList();
         return view('backend.equipment.types.create', compact('types'));
     }
 
@@ -57,7 +59,6 @@ class EquipmentTypeController extends Controller
             $type = new EquipmentType($data);
             $type->save();
             return redirect()->route('admin.equipment.types.index')->with('Success', 'EquipmentType was created !');
-
         } catch (\Exception $ex) {
             return abort(500, "Error 222");
         }
@@ -82,7 +83,7 @@ class EquipmentTypeController extends Controller
      */
     public function edit(EquipmentType $equipmentType)
     {
-        $types = EquipmentType::pluck('title', 'id');
+        $types =  EquipmentType::getFullTypeList();
         return view('backend.equipment.types.edit', compact('equipmentType', 'types'));
     }
 
@@ -105,12 +106,11 @@ class EquipmentTypeController extends Controller
 
         try {
             if ($request->thumb != null) {
-                $data['thumb'] = $this->uploadThumb($equipmentType->thumbURL(), $request->thumb, "equipment_types");
+                $data['thumb'] = $this->uploadThumb($equipmentType->thumb, $request->thumb, "equipment_types");
             }
 
             $equipmentType->update($data);
             return redirect()->route('admin.equipment.types.index')->with('Success', 'EquipmentType was updated !');
-
         } catch (\Exception $ex) {
             return abort(500);
         }
@@ -137,19 +137,21 @@ class EquipmentTypeController extends Controller
     {
         try {
             // Delete the thumbnail form the file system
-            $this->deleteThumb($equipmentType->thumbURL());
+            if ($equipmentType->thumb != null) {
+                $this->deleteThumb($equipmentType->thumb);
+            }
 
             $equipmentType->delete();
             return redirect()->route('admin.equipment.types.index')->with('Success', 'EquipmentType was deleted !');
-
         } catch (\Exception $ex) {
+            dd($ex);
             return abort(500);
         }
     }
 
     private function deleteThumb($currentURL)
     {
-        if ($currentURL != null) {
+        if ($currentURL != null && $currentURL != config('constants.frontend.dummy_thumb')) {
             $oldImage = public_path($currentURL);
             if (File::exists($oldImage)) unlink($oldImage);
         }
